+++ src/plugins/utils/utils-base64.ts
@@ -34,3 +34,16 @@ export function arrayBufferToBase64(buffer: ArrayBuffer) {
     return btoa(binary);
 }
 
+
+/**
+ * @link https://stackoverflow.com/a/21797381
+ */
+export function base64ToArrayBuffer(base64: string): ArrayBuffer {
+    const binary_string = atob(base64);
+    const len = binary_string.length;
+    const bytes = new Uint8Array(len);
+    for (let i = 0; i < len; i++) {
+        bytes[i] = binary_string.charCodeAt(i);
+    }
+    return bytes.buffer;
+}